requirements.txt
flask
openai
gunicorn
pandas
openpyxl
python-docx